+++ R/chrome.R
@@ -94,11 +94,14 @@ chrome_print = function(
     '--headless', '--no-first-run', '--no-default-browser-check', '--hide-scrollbars'
   ))
 
-  debug_port = servr::random_port()
+  debug_port = servr::random_port(NULL)
+  log_file = if (getOption('pagedown.chrome.log', FALSE)) {
+    sprintf('chrome-stderr-%s.log', format(Sys.time(), "%Y-%m-%d_%H-%M-%S"))
+  }
   ps = processx::process$new(browser, c(
     paste0('--remote-debugging-port=', debug_port),
     paste0('--user-data-dir=', work_dir), extra_args
-  ))
+  ), stderr = log_file)
   kill_chrome = function(...) {
     if (verbose >= 1) message('Closing browser')
     if (ps$is_alive()) ps$kill()
@@ -153,7 +156,7 @@ chrome_print = function(
       )
       svr = servr::httd(
         dirname(url), daemon = TRUE, browser = FALSE, verbose = verbose >= 1,
-        port = servr::random_port(), initpath = httpuv::encodeURIComponent(basename(url))
+        port = servr::random_port(NULL), initpath = httpuv::encodeURIComponent(basename(url))
       )
       stop_server = function(...) {
         if (verbose >= 1) message('Closing local webserver')