+++ NEWS.md
@@ -1,3 +1,21 @@
+# CHANGES IN pagedown VERSION 0.16
+
+## NEW FEATURES
+
+- Added support for long tables, i.e., repeat the table header when a page breaks the table (thanks, @felipecrp and @jdbarillas, #250 and #162).
+
+## MINOR CHANGES
+
+- Document that the `html_letter()` output format modifies the default value of the `fig_caption` parameter and allow the user to change it (thanks, @NewGraphEnvironment, #234).
+
+- Created an option `pagedown.chrome.log` to ease `chrome_print()` debugging: when set to `TRUE`, the logs of Chrome are saved in the working directory.
+
+- Added support for Sass files with `.scss` and `.sass` extension (thanks, @felipecrp, #244).
+
+## BUG FIXES
+
+- Fix an issue with `loft.lua` filter following changes in Pandoc 2.16 (#249).
+
 # CHANGES IN pagedown VERSION 0.15
 
 ## NEW FEATURES